refactor(temporal): optimization of jiff extension structure
