+++ modules/llrt_temporal/src/utils/span/total.rs
