+++ modules/llrt_temporal/src/utils/time/mod.rs
@@ -1,5 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
+pub mod round;
+
 use jiff::civil::Time;
 use llrt_utils::result::ResultExt;
 use rquickjs::{Ctx, Object, Result, Value};
