+++ modules/llrt_temporal/src/utils/timestamp/mod.rs
@@ -1,3 +1,3 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-pub mod span;
+pub mod round;
